+++ cpsycle/ui/src/uiscroller.c
@@ -85,8 +85,8 @@ void psy_ui_scroller_onsize(psy_ui_Scroller* self, psy_ui_Component* sender, psy
 void psy_ui_scroller_horizontal_onchanged(psy_ui_Scroller* self, psy_ui_ScrollBar* sender)
 {
 	assert(self->client);
-	int iPos;
-	int nPos;
+	intptr_t iPos;
+	intptr_t nPos;
 
 	iPos = self->client->scroll.x / self->client->scrollstepx;
 	nPos = psy_ui_scrollbar_position(sender);
@@ -98,8 +98,8 @@ void psy_ui_scroller_horizontal_onchanged(psy_ui_Scroller* self, psy_ui_ScrollBa
 void psy_ui_scroller_vertical_onchanged(psy_ui_Scroller* self, psy_ui_ScrollBar* sender)
 {
 	assert(self->client);
-	int iPos;
-	int nPos;
+	intptr_t iPos;
+	intptr_t nPos;
 
 	iPos = self->client->scroll.y / self->client->scrollstepy;
 	nPos = psy_ui_scrollbar_position(sender);
@@ -110,7 +110,7 @@ void psy_ui_scroller_vertical_onchanged(psy_ui_Scroller* self, psy_ui_ScrollBar*
 
 void psy_ui_scroller_onscroll(psy_ui_Scroller* self, psy_ui_Component* sender)
 {
-	int nPos;
+	intptr_t nPos;
 
 	nPos = self->client->scroll.y / self->client->scrollstepy;
 	psy_ui_scrollbar_setthumbposition(&self->vscroll, nPos);
